+++ src/main/java/com/brothers/premasiri/asset/employee/dao/EmployeeDao.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface EmployeeDao extends JpaRepository<Employee, Integer> {
+public interface EmployeeDao extends JpaRepository<Employee, Long> {
     Employee findFirstByOrderByIdDesc();
 
 }
