+++ src/main/java/com/brothers/premasiri/asset/supplier/dao/SupplierDao.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SupplierDao extends JpaRepository<Supplier, Integer> {
+public interface SupplierDao extends JpaRepository<Supplier, Long> {
 }
