+++ src/main/java/com/brothers/premasiri/asset/supplier/service/SupplierService.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 @Service
-public class SupplierService implements AbstractService<Supplier, Integer > {
+public class SupplierService implements AbstractService<Supplier, Long > {
 
 
 
@@ -16,7 +16,7 @@ public class SupplierService implements AbstractService<Supplier, Integer > {
     }
 
 
-    public Supplier findById(Integer id) {
+    public Supplier findById(Long id) {
         return null;
     }
 
@@ -26,7 +26,7 @@ public class SupplierService implements AbstractService<Supplier, Integer > {
     }
 
 
-    public boolean delete(Integer id) {
+    public boolean delete(Long id) {
         return false;
     }
